.gitlab-ci: check setup_repositories.sh with shellcheck
This is a new script, which has no shellcheck violations, so start
checking it to catch new failures, warnings and regressions.

The other, existing scripts are not checked, because they are not
passing yet.

Once they are passing, the shellcheck ci action should be set up
to check all the scripts.

Change-Id: Id9c7ffdd7b0cda01ee41b14f123c1d2a2176c614
Signed-off-by: Alexander Martinz <amartinz@shiftphones.com>