Drop support for building 16.04 images and default to 20.04
Ubuntu 16.04-based Ubuntu Touch has been long deprecated and
unmaintained, however the scripts may still produce an old structure
device tarball for 16.04, which is confusing for people porting
to new devices.

Drop support for 16.04 completely and keep 20.04 as the default
since we are not making new builds and don't expect new device ports
for 16.04.